Rename parser module to cli
